test: CarRepositoryTest와 CarDetailsRepositoryTest를 분리
